functionality to add location from map to form added
