+++ lib/src/providers/coordinates_provider.dart
@@ -2,16 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 
 class CoordinatesProvider extends ChangeNotifier {
-  int? giftInviteFriend;
   Marker? point;
+  String? name;
+  double? longitude;
+  double? latitude;
 
   CoordinatesProvider(BuildContext context) {
     //init
   }
 
   void addPoint(LatLng pos) {
-    print('ADD POINT ----------------');
-
     point = Marker(
         markerId: const MarkerId('point'),
         infoWindow:
@@ -20,13 +20,30 @@ class CoordinatesProvider extends ChangeNotifier {
           BitmapDescriptor.hueRose,
         ),
         position: pos);
-    notifyListeners();
 
-    print(point);
+    longitude = point!.position.longitude;
+    latitude = point!.position.latitude;
+
+    notifyListeners();
   }
 
   void clearPoint() {
     point = null;
+    notifyListeners();
+  }
+
+  void setName(String text) {
+    name = text;
+    notifyListeners();
+  }
+
+  void setLongitude(String long) {
+    longitude = double.parse(long);
+    notifyListeners();
+  }
+
+  void setLatitude(String lat) {
+    latitude = double.parse(lat);
     notifyListeners();
   }
 }
